split action code and make get request
