+++ src/store/cart-actions.js
@@ -1,0 +1,79 @@
+import { uiActions } from "./ui-slice";
+import { cartActions } from "./cart-slice";
+
+export const fetchCartData = () => {
+  return async (dispatch) => {
+    const fetchData = async () => {
+      const response = await fetch(
+        "https://react-hooks-fe144-default-rtdb.asia-southeast1.firebasedatabase.app/cart.json"
+      );
+
+      if (!response.ok) {
+        throw new Error("Loading fail.");
+      }
+
+      const data = await response.json();
+
+      return data;
+    };
+
+    try {
+      const cartData = await fetchData();
+      dispatch(cartActions.replaceCart(cartData));
+    } catch (error) {
+      dispatch(
+        uiActions.showNotification({
+          status: "error",
+          title: "Error",
+          message: "Fetching Data Fail.",
+        })
+      );
+    }
+  };
+};
+
+export const sendCartData = (cart) => {
+  return async (dispatch) => {
+    dispatch(
+      uiActions.showNotification({
+        status: "pending",
+        title: "Sending...",
+        message: "Sending cart data",
+      })
+    );
+
+    const sendRequest = async () => {
+      const response = await fetch(
+        "https://react-hooks-fe144-default-rtdb.asia-southeast1.firebasedatabase.app/cart.json",
+        {
+          method: "PUT",
+          body: JSON.stringify(cart),
+        }
+      );
+
+      if (!response.ok) {
+        throw new Error("Sending cart data fail.");
+      }
+    };
+
+    try {
+      await sendRequest();
+
+      dispatch(
+        uiActions.showNotification({
+          status: "success",
+          title: "Success",
+          message: "Send cart data successfully,",
+        })
+      );
+    } catch (error) {
+      dispatch(
+        uiActions.showNotification({
+          status: "error",
+          title: "error",
+          message: "Sending cart data failed.",
+        })
+      );
+    }
+  };
+};
